+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## v1.1.0 (unreleased)
 * Minimum requirements now Elixir v1.4 and OTP 19 (Kadabra bumped to `v0.3.0`)
-* `APNS.Config.config/1` and `ADM.Config.config/1` renamed to
-  `APNS.Config.new/1` and `ADM.Config.new/1`, respectively
+
+### APNS
+* `APNS.Config.config/1` renamed to `APNS.Config.new/1`
+* `APNS.push/2` without an `:on_response` callback now returns a
+  `%Pigeon.APNS.NotificationResponse{}`
+
+### ADM
+* `ADM.Config.config/1` renamed to `ADM.Config.new/1`
 
 ## v1.0.2
 * Fixed FCM infinite `GOAWAY session_timed_out` loop